refactor: Delete `-dialog` from directory name
