+++ frontend/src/app/app.module.ts
@@ -92,8 +92,8 @@ import { GuacamoleDialogComponent } from './sessions/user-sessions-wrapper/activ
 import { FileBrowserDialogComponent } from './sessions/user-sessions-wrapper/active-sessions/file-browser-dialog/file-browser-dialog.component';
 import { FileExistsDialogComponent } from './sessions/user-sessions-wrapper/active-sessions/file-browser-dialog/file-exists-dialog/file-exists-dialog.component';
 import { CreatePersistentSessionComponent } from './sessions/user-sessions-wrapper/create-session/create-persistent-session/create-persistent-session.component';
-import { CreateReadonlyModelOptionsComponent } from './sessions/user-sessions-wrapper/create-session/create-readonly-session-dialog/create-readonly-model-options/create-readonly-model-options.component';
-import { CreateReadonlySessionDialogComponent } from './sessions/user-sessions-wrapper/create-session/create-readonly-session-dialog/create-readonly-session-dialog.component';
+import { CreateReadonlyModelOptionsComponent } from './sessions/user-sessions-wrapper/create-session/create-readonly-session/create-readonly-model-options/create-readonly-model-options.component';
+import { CreateReadonlySessionDialogComponent } from './sessions/user-sessions-wrapper/create-session/create-readonly-session/create-readonly-session-dialog.component';
 import { UserSessionsWrapperComponent } from './sessions/user-sessions-wrapper/user-sessions-wrapper.component';
 import { AlertSettingsComponent } from './settings/core/alert-settings/alert-settings.component';
 import { ToolDeletionDialogComponent } from './settings/core/tools-settings/tool-details/tool-deletion-dialog/tool-deletion-dialog.component';
