+++ frontend/src/app/projects/project-detail/model-overview/model-overview.component.ts
@@ -13,7 +13,7 @@ import {
 import { ProjectUserService } from 'src/app/projects/project-detail/project-users/service/project-user.service';
 import { UserService } from 'src/app/services/user/user.service';
 import { SessionService } from 'src/app/sessions/service/session.service';
-import { CreateReadonlySessionDialogComponent } from 'src/app/sessions/user-sessions-wrapper/create-session/create-readonly-session-dialog/create-readonly-session-dialog.component';
+import { CreateReadonlySessionDialogComponent } from 'src/app/sessions/user-sessions-wrapper/create-session/create-readonly-session/create-readonly-session-dialog.component';
 import { TriggerPipelineComponent } from '../../models/backup-settings/trigger-pipeline/trigger-pipeline.component';
 import { ProjectService } from '../../service/project.service';
