+++ frontend/src/app/sessions/user-sessions-wrapper/create-session/create-readonly-session/create-readonly-model-options/create-readonly-model-options.component.html
