+++ frontend/src/app/sessions/user-sessions-wrapper/create-session/create-readonly-session/create-readonly-session-dialog.component.ts
